wells are done
