+++ app/src/components/map/LayerLoaderHelpers/DataBCRenderLayer.tsx
@@ -38,7 +38,6 @@ export const DataBCLayer = (props) => {
           simplifyPercentage={props.simplifyPercentage}
           opacity={props.opacity}
           dataBCLayerName={props.bcgw_code}
-          setClosestWells={props.setClosestWells}
         />
       );
     case LayerMode.VectorTilesOnline:
@@ -54,7 +53,6 @@ export const DataBCLayer = (props) => {
           simplifyPercentage={props.simplifyPercentage}
           opacity={props.opacity}
           dataBCLayerName={props.bcgw_code}
-          setClosestWells={props.setClosestWells}
         />
       );
     default:
